fix(mrc): move array excluded columns in constant file
ref: #MANAGER-17732
