+++ packages/manager-react-components/src/components/datagrid/datagrid-topbar.component.tsx
@@ -16,8 +16,10 @@ import { FilterWithLabel } from '../filters/interface';
 import { FilterAdd, FilterList } from '../filters';
 import { ColumnFilter } from '../filters/filter-add.component';
 import './translations';
-import { VisibilityManagement } from './visibility/visibility-management.component';
-import { ColumnsVisibility } from './datagrid.constants';
+import {
+  ColumnsVisibility,
+  VisibilityManagement,
+} from './visibility/visibility-management.component';
 
 type ColumnFilterProps = {
   key: string;
